Day 24
Part two is rather slow because of the cloning. It would be faster
to have 2 mutable world states but that is a bit tedious in rust
ownership land. I didn't come up with a good abstraction for the
layer adjacent checks so it's rather verbose clone of part 1
neighbour function.