Day 18
This was running rather slow with a bruteforce BFS so I had to
run it with --release flag to make it about a second or two.
Part 2 was pretty hard but it seems like ignoring doors you cannot
open and passing through them still gets the best solution. I think
there could be a case were this is not shortest path.